Rename git to git-import
