+++ src/webview/git-import/gitImportLoader.ts
@@ -6,12 +6,12 @@ import * as vscode from 'vscode';
 import * as path from 'path';
 import * as fs from 'fs';
 import { ExtensionID } from '../../util/constants';
-import { GitProvider } from '../../git/types/git';
-import { getGitService } from '../../git/services';
-import { DetectedServiceData, DetectedStrategy, detectImportStrategies } from '../../git/utils';
+import { GitProvider } from '../../git-import/types/git';
+import { getGitService } from '../../git-import/services';
+import { DetectedServiceData, DetectedStrategy, detectImportStrategies } from '../../git-import/utils';
 import { ComponentTypesView } from '../../registriesView';
 import { ComponentTypeDescription } from '../../odo/componentType';
-import { Response } from '../../git/types';
+import { Response } from '../../git-import/types';
 import { Uri, workspace } from 'vscode';
 import { Component } from '../../openshift/component';
 import OpenShiftItem from '../../openshift/openshiftItem';
